feat(utils): export utils functions to clone types
